Reorganized some stuff at the topmost levels to open the Wireworld project up a bit to applications that might instantiate it.
Added a separate timer for screen refreshing during overdrive. It's affected by whether the player is active or not.

Made the API a dynamic subclass of EventDispatcher. Perfect.

The Controller now discovers the File class when it initializes, if it is in the trusted application domain of an AIR app.