Added FontSet system to distribute fonts across the View. Lot of good that did me so far.
Came to my fucking senses and removed TLFTextField from the project. I'll be using TextLines for most things, rolling my own interactive text system for input text. The HTML will still be rendered in the dialogs with a TextField, but it'll use an embedded font to scale nicely.

Fixed readme markup, added about markup.

Moved the ImageSaver to the View and disconnected the Controller further from the bridge.